requirements.txt
glfw==2.8.0
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
